+++ src/main/java/xyz/foolcat/eve/evehelper/shared/util/UserUtil.java
@@ -20,7 +20,7 @@ public class UserUtil {
         }
         Object principal = authentication.getPrincipal();
         if (principal instanceof SysUser) {
-            return ((SysUser) principal).getId();
+            return ((SysUser) principal).getId().intValue();
         }
         return ((Long) principal).intValue();
     }
